+++ frontend/src/components/FillProfileForm.tsx
@@ -5,20 +5,25 @@ import { useUserStore } from "../store/useUserStore";
 import { useNavigate } from "react-router-dom";
 import Select from "react-select";
 import GenderSelect from "./GenderSelect";
+import SexualPreferenceSelect from "./SexualPreferenceSelect";
+import BioInput from "./BioInput";
+import InterestsSelect from "./InterestsSelect";
 
 function FillProfileForm() {
-  const navigate = useNavigate();
-
-  const { authUser, checkAuth } = useAuthStore();
-
-  const { loading, updateProfile } = useUserStore();
   const [gender, setGender] = useState("");
   const [sexualPreference, setSexualPreference] = useState("bisexual");
   const [bio, setBio] = useState("");
-  const [interests, setInterests] = useState<any>([]);
+  const [interests, setInterests] = useState<
+    { value: string; label: string }[]
+  >([]);
   const [pictures, setPictures] = useState<string[]>([]);
 
   const fileInputRef = useRef<HTMLInputElement>(null);
+
+  const navigate = useNavigate();
+
+  const { authUser, checkAuth } = useAuthStore();
+  const { loading, updateProfile } = useUserStore();
 
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
@@ -30,7 +35,7 @@ function FillProfileForm() {
         gender,
         sexual_preferences: sexualPreference,
         bio,
-        interests: interests.map((interest: any) => interest.value),
+        interests: interests,
         profile_picture: profilePicture,
         additional_pictures: pictures.slice(1),
       });
@@ -74,153 +79,21 @@ function FillProfileForm() {
     { value: "music", label: "#music" },
   ];
 
-  const controlStyles = {
-    base: "border rounded-lg bg-white hover:cursor-pointer",
-    focus: "ring-1 ring-purple-700",
-    nonFocus: "border-gray-300 hover:border-gray-400",
-  };
-  const placeholderStyles = "text-gray-500 pl-1 py-0.5 text-sm";
-  const selectInputStyles = "pl-1 py-0.5";
-  const valueContainerStyles = "p-1 gap-1";
-  const singleValueStyles = "leading-7 ml-1";
-  const multiValueStyles =
-    "bg-gray-100 rounded items-center py-0.5 pl-2 pr-1 gap-1.5";
-  const multiValueLabelStyles = "leading-6 py-0.5";
-  const multiValueRemoveStyles =
-    "border border-gray-200 bg-white hover:bg-red-50 hover:text-red-800 text-gray-500 hover:border-red-300 rounded-md";
-  const indicatorsContainerStyles = "p-1 gap-1";
-  const clearIndicatorStyles =
-    "text-gray-500 p-1 rounded-md hover:bg-red-50 hover:text-red-800";
-  const indicatorSeparatorStyles = "bg-gray-300";
-  const dropdownIndicatorStyles =
-    "p-1 hover:bg-gray-100 text-gray-500 rounded-md hover:text-black";
-  const menuStyles = "p-1 mt-2 border border-purple-200 bg-white rounded-lg";
-  const groupHeadingStyles = "ml-3 mt-2 mb-1 text-gray-500 text-sm";
-  const optionStyles = {
-    base: "hover:cursor-pointer px-3 py-2 rounded",
-    focus: "bg-gray-100 active:bg-gray-200",
-    selected:
-      "after:content-['✔'] after:ml-2 after:text-green-500 text-gray-500",
-  };
-  const noOptionsMessageStyles =
-    "text-gray-500 p-2 bg-gray-50 border border-dashed border-gray-200 rounded-sm";
-
   return (
     <div className="mt-8 sm:mx-auto sm:w-full sm:max-w-md">
       <div className="bg-white py-8 px-4 shadow sm:rounded-lg sm:px-10 border border-gray-200">
         <form onSubmit={handleSubmit} className="space-y-6">
-          {/* GENDER */}
           <GenderSelect gender={gender} setGender={setGender} />
-          {/* GENDER PREFERENCE */}
-          {/* <div>
-          <span className="block text-sm font-medium text-gray-700 mb-2">
-            Gender Preference
-          </span>
-          <div className="flex space-x-4">
-            {["heterosexual", "homosexual", "bisexual"].map((option) => (
-              <label key={option} className="inline-flex items-center">
-                <input
-                  type="checkbox"
-                  style={{
-                    accentColor: "#7E22CE",
-                  }}
-                  className="form-checkbox"
-                  checked={
-                    sexualPreference.toLowerCase() ===
-                    option.toLowerCase()
-                  }
-                  onChange={() =>
-                    setSexualPreference(option.toLowerCase())
-                  }
-                />
-                <span className="ml-2">{option}</span>
-              </label>
-            ))}
-          </div>
-        </div> */}
-          {/* BIO */}
-          <div>
-            <label
-              htmlFor="bio"
-              className="block text-sm font-medium text-gray-700 mb-2"
-            >
-              Bio
-            </label>
-            <div className="mt-1">
-              <textarea
-                id="bio"
-                name="bio"
-                rows={3}
-                value={bio}
-                onChange={(e) => setBio(e.target.value)}
-                className="appearance-none block w-full px-3 py-2 border border-gray-300 rounded-md shadow-sm
-          placeholder-gray-400 focus:outline-none focus:ring-purple-500 focus:border-purple-500 sm:text-sm"
-              />
-            </div>
-          </div>
-          {/* INTERESTS */}
-          <div>
-            <span className="block text-sm font-medium text-gray-700 mb-2">
-              Interests
-            </span>
-            <div>
-              <Select
-                isMulti
-                closeMenuOnSelect={false}
-                hideSelectedOptions={false}
-                unstyled
-                styles={{
-                  input: (base) => ({
-                    ...base,
-                    "input:focus": {
-                      boxShadow: "none",
-                    },
-                  }),
-                  // On mobile, the label will truncate automatically, so we want to
-                  // override that behaviour.
-                  multiValueLabel: (base) => ({
-                    ...base,
-                    whiteSpace: "normal",
-                    overflow: "visible",
-                  }),
-                  control: (base) => ({
-                    ...base,
-                    transition: "none",
-                  }),
-                }}
-                classNames={{
-                  control: ({ isFocused }) =>
-                    clsx(
-                      isFocused ? controlStyles.focus : controlStyles.nonFocus,
-                      controlStyles.base
-                    ),
-                  placeholder: () => placeholderStyles,
-                  input: () => selectInputStyles,
-                  valueContainer: () => valueContainerStyles,
-                  singleValue: () => singleValueStyles,
-                  multiValue: () => multiValueStyles,
-                  multiValueLabel: () => multiValueLabelStyles,
-                  multiValueRemove: () => multiValueRemoveStyles,
-                  indicatorsContainer: () => indicatorsContainerStyles,
-                  clearIndicator: () => clearIndicatorStyles,
-                  indicatorSeparator: () => indicatorSeparatorStyles,
-                  dropdownIndicator: () => dropdownIndicatorStyles,
-                  menu: () => menuStyles,
-                  groupHeading: () => groupHeadingStyles,
-                  option: ({ isFocused, isSelected }) =>
-                    clsx(
-                      isFocused && optionStyles.focus,
-                      isSelected && optionStyles.selected,
-                      optionStyles.base
-                    ),
-                  noOptionsMessage: () => noOptionsMessageStyles,
-                }}
-                value={interests}
-                onChange={setInterests}
-                options={options}
-              />
-            </div>
-          </div>
+          {/* <SexualPreferenceSelect
+            sexualPreference={sexualPreference}
+            setSexualPreference={setSexualPreference}
+          /> */}
+          <BioInput bio={bio} setBio={setBio} />
+          <InterestsSelect
+            interests={interests}
+            setInterests={setInterests}
+            options={options}
+          />
           {/* PICTURES */}
           <div>
             <label className="block text-sm font-medium text-gray-700">
